requirements.txt
# Core dependencies
streamlit>=1.28.0
google-generativeai>=0.3.0
arxiv>=1.4.0

# PDF processing
PyMuPDF>=1.23.0
pymupdf4llm>=0.0.5

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Text processing and NLP
nltk>=3.8.0
textstat>=0.7.0
python-docx>=0.8.11

# Utilities
python-dotenv>=1.0.0
pathlib>=1.0.1
logging>=0.4.9.6
requests>=2.31.0

# Optional: Enhanced features
scienceplots>=2.1.0  # For scientific plotting
wordcloud>=1.9.0     # For text visualization
spacy>=3.6.0         # For advanced NLP (optional)

# Development dependencies (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0